Reload root cid (#53)
This PR adds the ability for the dapp to automatically reload chapters. If there is an update to the chapter the player is currently in, the chapter is updated and a dialog is displayed. If not, the process is totally transparent to the user.
This PR reintroduces also the variable fuckFuckFuckFuckFuck in case the game state is f*cked. Ref: #52 (comment)

Demo:
    I'm a player at chapter 0.
    The game masters push an update to chapter 0 because there was an issue of some kind.
    The dapp automatically pulls the new changes and shows a dialog to notify me I should read the chapter again because it has been updated.
    When I click to acknowledge the update, the dialog disappears.